make requested changes
